refactor : Category 연관관계 재설정 및 User 탈퇴 시간 업데이트 개선
